+++ config.php
@@ -4,7 +4,9 @@ namespace MLL\RectorConfig;
 
 use Rector\Config\RectorConfig;
 
-/** Configure rector with PHP rules. */
+/**
+ * Configure rector with PHP rules.
+ */
 function config(RectorConfig $rectorConfig): void
 {
     // Use ArrayToFirstClassCallableRector in Rector 2.x, FirstClassCallableRector in 1.x
@@ -18,10 +20,13 @@ function config(RectorConfig $rectorConfig): void
     $rectorConfig->rule(\MLL\RectorConfig\Rector\IfThrowToCoalesceThrowRector::class);
 }
 
-/** Configure rector with Laravel rules. */
+/**
+ * Configure rector with Laravel rules.
+ */
 function laravel(RectorConfig $rectorConfig): void
 {
     config($rectorConfig);
+    $rectorConfig->rule(\MLL\RectorConfig\Rector\PrimaryKeyColumnToIdAliasRector::class);
     $rectorConfig->ruleWithConfiguration(\Rector\Transform\Rector\FuncCall\FuncCallToNewRector::class, [
         'collect' => \Illuminate\Support\Collection::class,
     ]);